Fixed a error when the video file was without sound

+++ VVC_GUI_Encoder.py
@@ -28,12 +28,16 @@ def EncodeButton():
 	global ii
 	os.system('ffmpeg_vvceasy.exe -i "'+filename+'" -q:a 0 -map a temp.wav')
 	os.system("exhale.exe c temp.wav temp.m4a") 
-	os.remove("temp.wav")
 	if passes.get() == "1 pass": 
 		os.system('ffmpeg_vvceasy.exe -y -i "'+filename+'" -pix_fmt yuv420p -f yuv4mpegpipe - | vvencapp.exe --y4m -i - --preset '+preset.get()+' -q '+quality.get()+' -o temp.266')
 	if passes.get() == "2 pass": 
 		os.system('ffmpeg_vvceasy.exe -y -i "'+filename+'" -pix_fmt yuv420p -f yuv4mpegpipe - | vvencapp.exe --y4m -i - --preset '+preset.get()+' --qpa 1 -p 2 -b '+qualitytwo.get()+'k -o temp.266')
-	os.system("mp4box.exe -add temp.266:fmt=VVC -add temp.m4a -new "+'"'+saveto.get()+'"')
+	if os.path.exists('temp.wav'):
+		os.remove("temp.wav")
+		os.system("mp4box.exe -add temp.266:fmt=VVC -add temp.m4a -new "+'"'+saveto.get()+'"')
+		os.remove("temp.m4a")
+	else:
+		os.system("mp4box.exe -add temp.266:fmt=VVC -new "+'"'+saveto.get()+'"')
 	os.remove("temp.266")
 	os.system("ffmpeg_vvceasy.exe -y -i "+'"'+saveto.get()+'"'+" -vf 'thumbnail' -frames:v 1  temp.jpg")
 	os.system("mp4box.exe -add "+'"'+saveto.get()+'"'+" -add temp.jpg -new "+'"'+saveto.get()+'"')
@@ -47,7 +51,6 @@ def EncodeButton():
 		wsize=int(imgtwo.size[0]*wpercent)
 		ii=ImageTk.PhotoImage(imgtwo.resize((wsize,200)))
 	canvas.create_image(180, 0, anchor='nw', image=ii)
-	os.remove("temp.m4a")
 	os.remove("temp.jpg")
 	fs2.config(text = 'Size(Mb): '+str(round(os.path.getsize(saveto.get())/1048576,2)))
 def btnClickFunctiontwo():
